requirements.txt
python-dotenv==0.14
discord.py==1.4
google-cloud-vision==1.0
Metaphone==0.6